runtime: review and update ACL perms for fns

+++ src/contract/dao/src/lib.rs
@@ -90,7 +90,7 @@ pub const DAO_CONTRACT_ZKAS_DAO_AUTH_MONEY_TRANSFER_NS: &str = "AuthMoneyTransfe
 pub const DAO_CONTRACT_ZKAS_DAO_AUTH_MONEY_TRANSFER_ENC_COIN_NS: &str = "AuthMoneyTransferEncCoin";
 
 /// Not allowed to make proposals using snapshots with block heights older than this depth
-pub const PROPOSAL_SNAPSHOT_CUTOFF_LIMIT: u64 = 100;
+pub const PROPOSAL_SNAPSHOT_CUTOFF_LIMIT: u32 = 100;
 
 // ANCHOR: dao-blockwindow
 const BLOCK_TIME: u64 = 90;
@@ -98,8 +98,8 @@ const SECS_IN_HOUR: u64 = 60 * 60;
 const WINDOW_TIME_HR: u64 = 4;
 
 /// Blockwindow from blockheight. Used for time limit on DAO proposals.
-pub fn blockwindow(height: u64) -> u64 {
-    let timestamp_secs = height * BLOCK_TIME;
+pub fn blockwindow(height: u32) -> u64 {
+    let timestamp_secs = height as u64 * BLOCK_TIME;
     timestamp_secs / (WINDOW_TIME_HR * SECS_IN_HOUR)
 }
 // ANCHOR_END: dao-blockwindow
